Fix scatterplot build.  Not perfect, but it should probably work and we should consolidate towards a replacement
