--- config/plugins/visualizations/scatterplot/.babelrc
@@ -1,11 +1,0 @@
-{
-    "presets": ["env"],
-    "plugins": [
-        [
-            "auto-import",
-            {
-                "declarations": [{ "default": "$", "path": "jquery" }, { "default": "jQuery", "path": "jquery" }]
-            }
-        ]
-    ]
-}
