+++ config/plugins/visualizations/scatterplot/package.json
@@ -10,8 +10,6 @@
     "author": "Carl Eberhard",
     "license": "0BSD",
     "dependencies": {
-        "babel-plugin-auto-import": "^0.7.1",
-        "babel-preset-env": "^1.6.1",
         "backbone": "^1.3.3",
         "bootstrap": "^3.3.7",
         "d3": "^3.5.17",
@@ -20,10 +18,9 @@
         "underscore": "^1.8.3"
     },
     "devDependencies": {
-        "@babel/core": "^7.24.6",
         "parcel": "2.11.0"
     },
     "scripts": {
-        "build": "parcel build src/scatterplot.js --dist-dir static/"
+        "build": "parcel build src/scatterplot.js --dist-dir static"
     }
 }
