+++ config/plugins/visualizations/scatterplot/src/jqglobals.js
@@ -1,2 +1,3 @@
-import $ from "jquery";
-window.jQuery = window.$ = window.jquery = $;
+import jquery from "jquery";
+
+export default (window.$ = window.jQuery = jquery);
