+++ config/plugins/visualizations/scatterplot/src/pagination.js
