+++ config/plugins/visualizations/scatterplot/src/peek-column-selector.js
